SAMA-48 replac edownlaod manager to class
