--- src/adapters/base/BaseAdapter.js
@@ -1,1 +1,0 @@
-export default class BaseAdapter {}
